Changed file structure for the css files renames static to styles and moved the css files out of ./static.css into ./styles/. In addition exported the Dashboard.jsx.
