+++ ui/src/styles/App.css
@@ -1,0 +1,11 @@
+.App {
+  text-align: center;
+}
+
+.data-area-div {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  text-align: left;
+}
